changed freshwater interface
