+++ animals/nene_goose.py
@@ -15,18 +15,6 @@ class NeneGoose(Animal, ITerrestrial, ISwimming, IWalking):
         self.__prey = { "grass", "plant", "leaves" }
 
 
-    @property
-    def prey(self):
-        return self.__prey
-
-    @property 
-    def species(self):
-        return self.__species
-
-    @property
-    def min_release_age(self):
-        return self.__min_release_age
-
     def feed(self, prey):
         if prey in self.__prey:
             print(f'The goose ate {prey} for a meal')
