+++ animals/river_dolphin.py
@@ -14,19 +14,6 @@ class RiverDolphin(Animal, IFreshwater, ISwimming, ISaltwater):
         self.age = age
         self.__prey = { "Trout", "Mackarel", "Salmon", "Sardine" }
 
-
-    @property
-    def prey(self):
-        return self.__prey
-
-    @property 
-    def species(self):
-        return self.__species
-
-    @property
-    def min_release_age(self):
-        return self.__min_release_age
-
     def feed(self, prey):
         if prey in self.__prey:
             print(f'The dolphin ate {prey} for a meal')
