+++ interfaces/organism/freshwater.py
@@ -4,5 +4,6 @@ class IFreshwater(IAquatic):
 
     def __init__(self):
         super().__init__()
-        self.tolerate_stagnant = False
         self.tolerate_fresh = True
+        self.tolerate_stagnant = False
+        self.tolerate_current = True
